refactor(team): cleanup pass — break circular import, drop dead code
Code review punch list:

- Extract PackageCardBody (+ CliHint, PackageEntryFields, ADOPT_INFO)
  into a new plugins/team/components/package-card.tsx so the
  agent-detail ↔ overview-tab cycle is broken (agent-detail imports
  OverviewTab; OverviewTab no longer needs to reach back into
  agent-detail).
- Drop unused `useCallback` from agent-detail.tsx imports.
- Drop unused PackageStateBadge + AdoptDialog imports from
  agent-detail.tsx (now only referenced by package-card).
- Drop dead WORKSPACE_FILES constant from openclaw-adapter.ts
  (never referenced anywhere).
- Strip stray double blank line in team-grid.tsx after
  ATTENTION_STATES.
- Fix misleading heartbeat copy: tooltip + empty state previously
  conflated HEARTBEAT.md (agent-maintained markdown narrative in the
  workspace) with the JSON status signal at
  ~/.bakin/heartbeats/{id}.json (written by bakin_exec_heartbeat).
  These are two distinct surfaces; the empty state now disambiguates.
